# Speed up compilation by 23%