--- a/Cargo.toml
+++ b/Cargo.toml
@@ -8,7 +8,8 @@ repository = "https://github.com/rust-lang/rustdoc-types"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = {version="1", features=["derive"]}
+serde = {version="1"}
+serde_derive = {version="1"}
 rustc-hash = {version="2", optional=true}
 
 [features]