Before this commit, serde_derive is built before serde. But serde does not depend on serde_derive, so that is not needed. Instead, build serde and serde_derive in parallel. That speeds up compilation by 23%.

Before:

    $ hyperfine --prepare 'cargo clean' 'cargo build'
    Benchmark 1: cargo build
      Time (mean ± σ):      4.979 s ±  0.044 s    [User: 4.887 s, System: 0.714 s]
      Range (min … max):    4.915 s …  5.068 s    10 runs

After:

    $ hyperfine --prepare 'cargo clean' 'cargo build'
    Benchmark 1: cargo build
      Time (mean ± σ):      3.821 s ±  0.065 s    [User: 5.083 s, System: 0.740 s]
      Range (min … max):    3.749 s …  3.955 s    10 runs