@aDotInTheVoid 2025-06-03
This will get blown away the next time `update.sh` is run. To avoid this, I'm going to submit a corresponding change to rust-lang/rust.

Also, we should use triagebot to add a comment for all PR's that touch `src/` as that's almost always wrong.