@aDotInTheVoid 2025-06-05
Published as d7f859a3d0206f3a68c8163f4d099c31d6a7f0fa (0.46.1), which also specifies that we need serde to be at least 1.0.186